move combined tree items to shortcuts logic
